+++ index.html
@@ -5,17 +5,19 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <title>Kamercrisis ||</title>
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" type="text/css" media="screen" href="main.css" />
-    <link rel="stylesheet" type="text/css" media="screen" href="navbar.css" />
-    <link rel="stylesheet" type="text/css" media="screen" href="hitmebutton.css" />
-    <link rel="stylesheet" type="text/css" media="screen" href="loginmodal.css" />
-    <link rel="shortcut icon" href="favicon(1).ico"/>
+    <link rel="stylesheet" type="text/css" media="screen" href="styles/main.css" />
+    <link rel="stylesheet" type="text/css" media="screen" href="styles/navbar.css" />
+    <link rel="stylesheet" type="text/css" media="screen" href="styles/hitmebutton.css" />
+    <link rel="stylesheet" type="text/css" media="screen" href="styles/loginmodal.css" />
+    <link rel="shortcut icon" href="assets/favicon(1).ico"/>
     <link href="https://fonts.googleapis.com/css?family=Quicksand" rel="stylesheet">
     <link href="https://fonts.googleapis.com/css?family=Roboto" rel="stylesheet">
     <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.6.1/css/all.css" integrity="sha384-gfdkjb5BdAXd+lj+gudLWI+BXq4IuLW5IT+brZEZsLFm++aCMlF1V92rMkPaX4PP" crossorigin="anonymous">
    <script src="https://ajax.aspnetcdn.com/ajax/jQuery/jquery-3.3.1.min.js"></script>
-    <script src="main.js"></script>
-    <script src="navbar.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/axios/0.18.0/axios.min.js"></script>
+    <script src="js/main.js"></script>
+    <script src="js/navbar.js"></script>
+    <script src="js/crudservice.js"></script>
 
 </head>
 
@@ -78,7 +80,7 @@
       <div class="container">
             <div class= "menu">
               <ul class="navbar-ul">
-                <li class="li-item1-navbar li-nav" >Home</li>
+                <li class="li-item1-navbar li-nav" onclick="getAllTodos()">Home</li>
                 <li class="li-item2-navbar li-nav" >About</li>
                 <li class="li-item3-navbar li-nav" >Events</li>
                 <li class="li-item4-navbar li-nav" >Search</li>
