+++ js/services.js
@@ -1,0 +1,12 @@
+const axios = require('./axios');
+
+axios.get(placeholder)
+.then((response)=>{
+    //do something with response
+    // response is a json file
+    let theinfoIamlookingfor = response['user'];
+    let myregister = $('#regist').value
+    if(theinfoIamlookingfor === myregister){
+        alert('you are already registered')
+    }
+})
