--- main.css
@@ -1,275 +1,0 @@
-
-body{
-    margin:0;
-    font-family: 'Quicksand', sans-serif;
-    /* margin:0px 180px 0px 180px; */
-}
-/* pageloader */
-.loader {
-    position: fixed;
-    z-index: 99;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background: white;
-    display: flex;
-    justify-content: center;
-    align-items: center; 
-}
-
-.loader > img {
-    width: 500px;
-}
-
-.loader.hidden {
-   animation: fadeOut 0.5s;
-   animation-fill-mode: forwards; 
-   animation-delay: 1s;
-}
-
-@keyframes fadeOut {
-    100% {
-        opacity: 0;
-        visibility: hidden;
-    }
-}
-
-/* end of pageloader */
-
-.container{
- margin:0px 180px 0px 180px; 
-}
-.container-header{
-margin:0px 180px 0px 180px;  
-}
-
-
-.h2-housing-project{
-    top: -9px;
-    position: relative;
-    display: initial;
-    margin: 5px;
-}
-header{
-
-    background-color: #005480;
-}
-/* ............................................................ */
-/* navbar  */
-.navbar-ul{
- 
-   font-size: 25px;
-   list-style-type: none;
-   text-align: center;
-   margin: 0;
-   padding: 0;
-   
-}
-nav{
-    display: inline-block;
-}
-.navbar-nav{
-    border: #005480 solid 1px;
-
-}
-.login-register-ul-right-top{
-    list-style-type: none;
-    padding: 4px;
-}
-.login-register-right-top-li{
-    margin-right:20px;
-    color:black;
-    float:right;
-    top: 56px; 
-    position:relative;
-}
-.login-register-right-top-li:hover{
-    border-bottom: black solid 1.5px;    
-    cursor: pointer;
-}
-.li-nav{
-    /* font-family: 'B612 Mono', monospace; */
-    display:block;
-    color:white;
-    padding: 4px;
-    float: left;
-    margin:0px 60px 0px 10px;
-}
-.li-nav:hover{
-    background-color: #003d5d;
- 
-    border-bottom-left-radius: 10px;
-    border-bottom-right-radius: 10px;
-    
-}
-/* inputsearchbar and form*/
-form{
-    margin: 0 auto; 
-    width:360px;
-}
-
-.position-absolute-for-input{
-   /* To center a search element */
-    margin: 255px 0px 0px 69px;
-    position: absolute;
-    left: 50%;
-    width: 900px;
-    margin-left: -450px;
-}   
-input{
-  
-    max-width:600px;
-    font-size: 20px;
-    padding: 15px;
-    margin: auto;
-    border: solid 1px #005480;
-    border-radius:3px;
-}   
-
-
-/* end of inputsearchbar and form */
-
-
-/* end of navbar */
-/* ,....................................,,,,,,,,,,,,,,,,,,,,,,, */
-
-/* /////////////////////////////////////////// */
-/* gridlayers */
-
-h1{
-  font-family: 'Quicksand', sans-serif;
-}
-.grid-container {
-    text-align: center;
-    margin: auto;
-    display: grid;
-    grid-template-columns: 1fr 1fr 1fr;
-    grid-template-rows: 1fr;
-    grid-template-areas: "1 2 3";
-  }
-.img-icon-grid{
-    margin-left:50px;
-    padding-top:40px;
-}
-
-/* hitmebutton */
-.center-button-hit-me-sign-up{
-    margin: 0 auto;
-    width: 200px;
-}
-
-/* endofhitmebutton */
-
-/* offersection */
-.img-responsive-offer-section{
-    width: 100%;
-    height: auto;
-    
-}
-.thumbnail-styling-picture:hover{
-    -webkit-box-shadow: 0px 2px 32px -7px rgba(0,0,0,0.75);
--moz-box-shadow: 0px 2px 32px -7px rgba(0,0,0,0.75);
-box-shadow: 0px 2px 32px -7px rgba(0,0,0,0.75);
-
-    
-}
-.thumbnail-styling-picture{
-    margin: 5px;
-    margin-bottom: 1.875rem;     
-    position: relative;
-    display: block;
-    border:solid 20px white;
-    border-radius:25px;
-    background-color: #fff;
-    color: #524c61;
-    text-decoration: none;
-    cursor: pointer;
-    -webkit-box-shadow: 0 1px 3px 0 rgba(37,32,31,.3);
-}
-.grid-container-offer-section{
-    display: grid;
-    grid-template-columns: 1fr 1fr 1fr;
-    grid-template-rows: 1fr 1fr  ;
-    grid-template-areas: "1 2" "3 4" "5 6";
-  }
-/* endofoffersection */
-/* footer */
-.grid-container-footer {
-    text-align: center;
-    color:white;
-    background-color: #005480;
-    display: grid;
-    grid-template-columns: 1fr 1fr 1fr 1fr;
-    grid-template-rows: 1fr 1fr;
-    grid-template-areas: "colum colum colum colum" "colum colum colum colum";
-    padding:20px 180px 0px 180px; 
-  }
-.icon-grid-footer{
-    font-size:25px;
-}
-/* end of footer */
-
-/* media queries */
-@media(max-width:1400px){
-    .container-resize-querie{
-        margin:0;
-    }
-}
-
-@media(max-width:800px){
-    .container-resize-querie{
-        margin:0px 180px 0px 180px; 
-    }
-    .grid-container{
-        display:block;
-    }
-    .grid-container-offer-section{
-        display:block;
-    }
-    .grid-container-footer{
-        display:block;
-        padding:20px;
-    }
-
-}
-@media(max-width:900px){
-    .container-header{
-        margin:0;
-    }
-}
-
-@media(max-width:500px){
-.login-register-ul-right-top{
-    float:right;
-    position: relative;
-    bottom: 30px;
-}
-.login-register-right-top-li{
-    float:none;
-    padding:4px;
-    
-}
-
-}
-
-@media(max-width:600px){
-    .container-resize-querie{
-        padding:0;
-        margin:60px;
-    }
-}
-/* end of media queries */
-.animate-bottom{
-    position:relative;
-    animation:animatebottom 1s}
-@keyframes animatebottom
-{from
-    {bottom:-300px;opacity:0}
-     to{bottom:0;opacity:1}
-    }
-
-    .image-header-img{
-        background-repeat: no-repeat;
-        background-size: 300px 100px;
-    }
